New load default option on Main Menu Modal, doesn't perform on startup
